pyproject.toml
[project]
name = "louisville-politics-tracker"
version = "0.1.0"
description = "Searchable dashboard of Louisville Metro Council + JCPS Board meetings."
requires-python = ">=3.10"
dependencies = [
    "civic-scraper>=0.2",
    "pdfminer.six>=20231228",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "pytest>=8",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["scrapers*"]
